backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
openai
diskcache
pyautogen
asyncio